requirements.txt
openai>=1.0.0
python-docx>=1.1.0
pypdf>=4.0.0
